api/requirements.txt
autopep8==1.5.4
Django==3.0.6
django-cors-headers==3.3.0
django-environ==0.4.5
djangorestframework==3.11.0
psycopg2-binary==2.8.5
